+++ framework/model/src/main/java/is/codion/framework/model/ForeignKeyConditionModel.java
@@ -20,6 +20,7 @@ package is.codion.framework.model;
 
 import is.codion.common.Operator;
 import is.codion.common.model.condition.ConditionModel;
+import is.codion.common.observer.Mutable;
 import is.codion.common.observer.Observer;
 import is.codion.common.state.State;
 import is.codion.common.value.Value;
@@ -97,7 +98,7 @@ public final class ForeignKeyConditionModel implements ConditionModel<Entity> {
 	}
 
 	@Override
-	public Value<Wildcard> wildcard() {
+	public Mutable<Wildcard> wildcard() {
 		return condition.wildcard();
 	}